scripts/requirements.txt
boto3>=1.34.84
toml
pyyaml
